fix(utilities): change structure to prevent circular dependencies
